+++ jsapp/js/components/processing/processingUtils.ts
@@ -1,5 +1,4 @@
 import type {LanguageCode} from 'js/components/languages/languagesStore';
-import type {ProcessingTabName} from 'js/components/processing/singleProcessingStore';
 
 type SupplementalPathPartsType = 'transcript' | 'translation' | 'qual';
 
@@ -54,42 +53,4 @@ export function getSupplementalPathParts(path: string): SupplementalPathParts {
   }
 
   return output;
-}
-
-/**
- * Checks if two processing paths differ only with the tab part and are
- * identical
- */
-export function isSamePathExceptTab(firstPath: string, secondPath: string) {
-  const firstPathArray = firstPath.split('/');
-  const secondPathArray = secondPath.split('/');
-  // Remove the tab part from both paths and compare
-  return (
-    firstPathArray.splice(7, 1).join('/') ===
-    secondPathArray.splice(7, 1).join('/')
-  );
-}
-
-interface ProcessingPathParts {
-  assetUid: string;
-  qpath: string;
-  submissionEditId: string;
-  tab: ProcessingTabName;
-}
-
-/**
- * For given processing path, returns all params
- */
-export function getProcessingPathParts(path: string): ProcessingPathParts {
-  const pathArray = path.split('/');
-
-  // We assume this will always be correct :fingers_crossed:
-  const pathTabPart = pathArray[7] as ProcessingTabName;
-
-  return {
-    assetUid: pathArray[2],
-    qpath: pathArray[5],
-    submissionEditId: pathArray[6],
-    tab: pathTabPart,
-  };
 }
